Move tests to UnitTests folder
