+++ tests/Zendesk/API/UnitTests/BasicTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Zendesk\API\LiveTests;
+namespace Zendesk\API\UnitTests;
 
 use Zendesk\API\HttpClient;
 use \Aeris\GuzzleHttpMock\Mock as GuzzleHttpMock;
