+++ tests/Zendesk/API/UnitTests/TicketCommentsTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Zendesk\API\LiveTests;
+namespace Zendesk\API\UnitTests;
 
 /**
  * Ticket Comments test class
