+++ tests/Zendesk/API/UnitTests/TicketsTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Zendesk\API\LiveTests;
+namespace Zendesk\API\UnitTests;
 
 
 /**
