+++ tests/Zendesk/API/UnitTests/UsersTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Zendesk\API\LiveTests;
+namespace Zendesk\API\UnitTests;
 
 use Zendesk\API\Resources\Users;
